pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm>=6.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[project]
name = "turep"
version = "0.1.0"
description = "Tumor-reactivity prediction for single-cell and spatial transcriptomics data."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Wendao Liu", email = "liuwd15@gmail.com"}
]
keywords = ["single-cell", "bioinformatics", "machine-learning", "vae", "focal-loss"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "torch>=1.12.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "scvi-tools>=1.3.0",
    "anndata>=0.8.0",
    "lightning>=2.0.0",
    "torchmetrics>=0.11.0",
    "platformdirs>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.950",
]

[project.urls]
Homepage = "https://github.com/liuwd15/turep"
Repository = "https://github.com/liuwd15/turep"
Documentation = "https://github.com/liuwd15/turep/blob/main/README.md"

[tool.black]
line-length = 100
target-version = ['py311', 'py312', 'py313']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
